+++ frontend/workflows/audit/tsconfig.json
@@ -1,30 +1,11 @@
 {
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
-    "declaration": true,
-    "esModuleInterop": true,
-    "importsNotUsedAsValues": "error",
-    "incremental": true,
-    "jsx": "react",
-    "lib": ["DOM", "ESNext"],
-    "moduleResolution": "node",
-    "noImplicitAny": false,
-    "noImplicitReturns": true,
     "outDir": "./dist",
-    "preserveConstEnums": true,
-    "rootDir": "./src",
-    "target": "ES2018",
-    "types": ["jest", "node", "long"],
+    "rootDir": "./src"
   },
   "include": [
     "src/**/*",
-  ],
-  "exclude": [
-    "node_modules",
-    "**/*.stories.ts*",
-    "**/*.stories.js*",
-    "**/*.test.js*",
-    "**/*.test.ts*",
-    "**/dist/*"
   ],
   "references": [
     {"path": "../../api"},
